fix(tests): migrate integration API blueprint tests to common API tests
